Todo esto es la integración de las tareas con lo que ya se tenía anteriormente. Además de eso, se corrigieron bugs que ocurrían en el online de la anterior versión. Actualmente solo falta el sabotaje y el botón de reuniones.

+++ core/src/main/java/com/amongus/core/model/Position.java
@@ -1,4 +1,3 @@
 package com.amongus.core.model;
 
-public record Position(int x, int y) {
-}
+public record Position(float x, float y) {}
